Added ability to override protofile root path. Added base page.
The base page is served from localhost:<your-port> and lists all of the RPCs
that are provided by the mock server. Clicking on any one of them will generate
a response and render it in the browser.

Added indents to handler response.

In practise, consumers shouldn't care and it make the responses easier to read as a human.

Changed to produce RPC responses even if unsupported type is encountered